requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra dev --output-file requirements-dev.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
boolean-py==5.0
    # via license-expression
build==1.2.2.post1
    # via pip-tools
cachecontrol==0.14.3
    # via pip-audit
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via pip-tools
cyclonedx-python-lib==9.1.0
    # via pip-audit
defusedxml==0.7.1
    # via py-serializable
filelock==3.18.0
    # via cachecontrol
flake8==7.3.0
    # via tsi-data-uploader (pyproject.toml)
google-api-core==2.25.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via tsi-data-uploader (pyproject.toml)
google-cloud-secret-manager==2.24.0
    # via tsi-data-uploader (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
greenlet==3.2.3
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.73.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via tsi-data-uploader (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via tsi-data-uploader (pyproject.toml)
license-expression==30.4.4
    # via cyclonedx-python-lib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.1
    # via cachecontrol
nest-asyncio==1.6.0
    # via tsi-data-uploader (pyproject.toml)
numpy==2.3.1
    # via pandas
opentelemetry-api==1.35.0
    # via google-cloud-logging
packageurl-python==0.17.3
    # via cyclonedx-python-lib
packaging==25.0
    # via
    #   build
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
pandas==2.3.1
    # via tsi-data-uploader (pyproject.toml)
tabulate
pip==25.2
    # via
    #   pip-api
    #   pip-tools
pip-api==0.0.34
    # via pip-audit
pip-audit==2.9.0
    # via tsi-data-uploader (pyproject.toml)
pip-requirements-parser==32.0.1
    # via pip-audit
pip-tools==7.4.1
    # via tsi-data-uploader (pyproject.toml)
platformdirs==4.3.8
    # via pip-audit
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==6.30.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.9
    # via tsi-data-uploader (pyproject.toml)
py-serializable==2.1.0
    # via cyclonedx-python-lib
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.14.0
    # via flake8
pydantic==2.11.7
    # via tsi-data-uploader (pyproject.toml)
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyparsing==3.2.3
    # via pip-requirements-parser
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   tsi-data-uploader (pyproject.toml)
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==1.1.0
    # via tsi-data-uploader (pyproject.toml)
pytest-mock==3.14.1
    # via tsi-data-uploader (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via tsi-data-uploader (pyproject.toml)
pytz==2025.2
    # via pandas
requests==2.32.4
    # via
    #   cachecontrol
    #   google-api-core
    #   pip-audit
rich==14.1.0
    # via pip-audit
rsa==4.9.1
    # via google-auth
ruff==0.12.7
    # via tsi-data-uploader (pyproject.toml)
setuptools==80.9.0
    # via pip-tools
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
sqlalchemy==2.0.41
    # via tsi-data-uploader (pyproject.toml)
toml==0.10.2
    # via pip-audit
tqdm==4.66.4
    # via tsi-data-uploader (pyproject.toml)
typing-extensions==4.14.1
    # via
    #   anyio
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
wheel==0.45.1
    # via pip-tools
zipp==3.23.0
    # via importlib-metadata
dbt-bigquery==1.10.1
    # pinned to avoid long pip resolver backtracking; update deliberately via uv/pip-compile
